ChatGPT generated Solidity code
This contract defines a Recipe struct that stores the name and ingredients of a single recipe. It also defines a RecipeBook contract that allows users to add new recipes and search for existing recipes based on a list of input ingredients. To use this contract, a user would first add new recipes using the addRecipe function, passing in the recipe name and a list of ingredients. Then, they could search for matching recipes using the searchRecipes function, passing in a list of ingredients that they have available. The function would return an array of Recipe structs that contain all of the input ingredients.